move adapter and view holder to same package
