+++ criminalintentsample/src/main/java/com/bignerdranch/android/criminalintent/CrimeChildViewHolder.java
@@ -4,7 +4,7 @@ import android.view.View;
 import android.widget.CheckBox;
 import android.widget.TextView;
 
-import com.bignerdranch.expandablerecyclerview.ViewHolder.ChildViewHolder;
+import com.bignerdranch.expandablerecyclerview.ChildViewHolder;
 
 /**
  * Created by ryanbrooks on 6/17/15.
