+++ criminalintentsample/src/main/java/com/bignerdranch/android/criminalintent/CrimeParentViewHolder.java
@@ -6,7 +6,7 @@ import android.view.View;
 import android.widget.ImageButton;
 import android.widget.TextView;
 
-import com.bignerdranch.expandablerecyclerview.ViewHolder.ParentViewHolder;
+import com.bignerdranch.expandablerecyclerview.ParentViewHolder;
 
 public class CrimeParentViewHolder extends ParentViewHolder {
     private static final float INITIAL_POSITION = 0.0f;
