+++ expandablerecyclerview/src/main/java/com/bignerdranch/expandablerecyclerview/ChildViewHolder.java
@@ -1,18 +1,13 @@
-package com.bignerdranch.expandablerecyclerview.ViewHolder;
+package com.bignerdranch.expandablerecyclerview;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
 
 /**
- * ViewHolder for a child list
- * item.
+ * ViewHolder for a child list item.
  * <p>
  * The user should extend this class and implement as they wish for their
  * child list item.
- *
- * @author Ryan Brooks
- * @version 1.0
- * @since 5/27/2015
  */
 public class ChildViewHolder extends RecyclerView.ViewHolder {
